Rough pass making silent classes always output; CSS classes optional
$use-html-classes outputs CSS classes for the .grid* classes only.

$use-all-html-classes outputs CSS classes for everything else.

Both must be set explicitly, there is no interaction between the two.

This happens by calling the setup mixins twice, once with each prefix.
It seems to work but hasn't been thoroughly tested. In particular, it's
hard to see what's going on when both variables are false i.e. no CSS
classes are output at all, but @extend seems to work fine. Note that the
first '%' pass creates and then @extends '%*' classes and the second '.'
pass creates and @extends '.*' classes. It was done like this simply
because it meant less editing, as we'll have to maintain this as a patch
for a while. Speaking of which, there's some whitespace changes that
shouldn't have been commited, we should probably redo this on top of
master.